Make the time stamp scale factor 64-bit.
It's calculated as the quotient of two 64-bit values, and is used only
in 64-bit calculations, so it should be made 64-bit as well.

(cherry picked from commit 2cac04a8d99d06da8eb2a0e12e9304300da6c93f)

+++ sf-pcapng.c
@@ -199,7 +199,7 @@ typedef enum {
 struct pcap_ng_if {
 	uint64_t tsresol;		/* time stamp resolution */
 	tstamp_scale_type_t scale_type;	/* how to scale */
-	u_int scale_factor;		/* time stamp scale factor for power-of-10 tsresol */
+	uint64_t scale_factor;		/* time stamp scale factor for power-of-10 tsresol */
 	uint64_t tsoffset;		/* time stamp offset */
 };
